Add an input placeholder "Tap here to type" on touchscreen devices
On touchscreen devices, if you blur focus from the primary input, it can be hard to figure out how to bring the keyboard back up. ("What's that > thing?")

So, on touchscreen devices, we'll add input placeholder text, "Tap here to type" and use CSS to hide that placeholder text if the input is currently focused.